runpod_worker/requirements.txt
# RunPod Worker Dependencies
runpod>=1.6.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0

# Core ML Dependencies
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0

# Additional Dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Image Processing
opencv-python-headless>=4.8.0
scikit-image>=0.21.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
psutil>=5.9.0
